+++ crates/task-impls/src/consensus/mod.rs
@@ -1,3 +1,5 @@
+#![cfg(not(feature = "dependency-tasks"))]
+
 use std::{collections::BTreeMap, sync::Arc};
 
 use anyhow::Result;
@@ -25,7 +27,6 @@ use hotshot_types::{
 use tokio::task::JoinHandle;
 use tracing::{debug, error, instrument, warn};
 use vbs::version::Version;
-#[cfg(not(feature = "dependency-tasks"))]
 use {
     crate::consensus::helpers::{
         handle_quorum_proposal_recv, handle_quorum_proposal_validated, publish_proposal_if_able,
@@ -41,9 +42,8 @@ use {
 };
 
 use crate::{
-    consensus::view_change::{update_view, DONT_SEND_VIEW_CHANGE_EVENT},
     events::{HotShotEvent, HotShotTaskCompleted},
-    helpers::{broadcast_event, cancel_task},
+    helpers::{broadcast_event, cancel_task, update_view, DONT_SEND_VIEW_CHANGE_EVENT},
     vote_collection::{
         create_vote_accumulator, AccumulatorInfo, HandleVoteEvent, VoteCollectionTaskState,
     },
@@ -51,9 +51,6 @@ use crate::{
 
 /// Helper functions to handle proposal-related functionality.
 pub(crate) mod helpers;
-
-/// Handles view-change related functionality.
-pub(crate) mod view_change;
 
 /// Alias for Optional type for Vote Collectors
 type VoteCollectorOption<TYPES, VOTE, CERT> = Option<VoteCollectionTaskState<TYPES, VOTE, CERT>>;
@@ -161,7 +158,6 @@ impl<TYPES: NodeType, I: NodeImplementation<TYPES>> ConsensusTaskState<TYPES, I>
     }
 
     /// Validate the VID disperse is correctly signed and has the correct share.
-    #[cfg(not(feature = "dependency-tasks"))]
     fn validate_disperse(&self, disperse: &Proposal<TYPES, VidDisperseShare<TYPES>>) -> bool {
         let view = disperse.data.view_number();
         let payload_commitment = disperse.data.payload_commitment;
@@ -206,7 +202,6 @@ impl<TYPES: NodeType, I: NodeImplementation<TYPES>> ConsensusTaskState<TYPES, I>
         true
     }
 
-    #[cfg(not(feature = "dependency-tasks"))]
     /// Publishes a proposal
     async fn publish_proposal(
         &mut self,
@@ -240,7 +235,6 @@ impl<TYPES: NodeType, I: NodeImplementation<TYPES>> ConsensusTaskState<TYPES, I>
 
     /// Spawn a vote task for the given view.  Will try to vote
     /// and emit a `QuorumVoteSend` event we should vote on the current proposal
-    #[cfg(not(feature = "dependency-tasks"))]
     async fn spawn_vote_task(
         &mut self,
         view: TYPES::Time,
@@ -285,10 +279,8 @@ impl<TYPES: NodeType, I: NodeImplementation<TYPES>> ConsensusTaskState<TYPES, I>
         event: Arc<HotShotEvent<TYPES>>,
         event_stream: Sender<Arc<HotShotEvent<TYPES>>>,
     ) {
-        #[cfg(not(feature = "dependency-tasks"))]
         let version = *self.version.read().await;
         match event.as_ref() {
-            #[cfg(not(feature = "dependency-tasks"))]
             HotShotEvent::QuorumProposalRecv(proposal, sender) => {
                 debug!("proposal recv view: {:?}", proposal.data.view_number());
                 match handle_quorum_proposal_recv(
@@ -309,7 +301,6 @@ impl<TYPES: NodeType, I: NodeImplementation<TYPES>> ConsensusTaskState<TYPES, I>
                     Err(e) => debug!("Failed to propose {e:#}"),
                 }
             }
-            #[cfg(not(feature = "dependency-tasks"))]
             HotShotEvent::QuorumProposalValidated(proposal, _) => {
                 debug!("proposal validated view: {:?}", proposal.view_number());
                 if let Err(e) =
@@ -397,7 +388,6 @@ impl<TYPES: NodeType, I: NodeImplementation<TYPES>> ConsensusTaskState<TYPES, I>
                     }
                 }
             }
-            #[cfg(not(feature = "dependency-tasks"))]
             HotShotEvent::QcFormed(cert) => match cert {
                 either::Right(qc) => {
                     self.proposal_cert = Some(ViewChangeEvidence::Timeout(qc.clone()));
@@ -448,7 +438,6 @@ impl<TYPES: NodeType, I: NodeImplementation<TYPES>> ConsensusTaskState<TYPES, I>
                     self.formed_upgrade_certificate = Some(cert.clone());
                 }
             }
-            #[cfg(not(feature = "dependency-tasks"))]
             HotShotEvent::DaCertificateRecv(cert) => {
                 debug!("DAC Received for view {}!", *cert.view_number);
                 let view = cert.view_number;
@@ -465,7 +454,6 @@ impl<TYPES: NodeType, I: NodeImplementation<TYPES>> ConsensusTaskState<TYPES, I>
                 }
                 self.spawn_vote_task(view, event_stream).await;
             }
-            #[cfg(not(feature = "dependency-tasks"))]
             HotShotEvent::VidShareRecv(disperse) => {
                 let view = disperse.data.view_number();
 
@@ -609,7 +597,6 @@ impl<TYPES: NodeType, I: NodeImplementation<TYPES>> ConsensusTaskState<TYPES, I>
                     consensus.metrics.number_of_timeouts_as_leader.add(1);
                 }
             }
-            #[cfg(not(feature = "dependency-tasks"))]
             HotShotEvent::SendPayloadCommitmentAndMetadata(
                 payload_commitment,
                 builder_commitment,
@@ -663,7 +650,6 @@ impl<TYPES: NodeType, I: NodeImplementation<TYPES>> ConsensusTaskState<TYPES, I>
                     }
                 }
             }
-            #[cfg(not(feature = "dependency-tasks"))]
             HotShotEvent::ViewSyncFinalizeCertificate2Recv(certificate) => {
                 if !certificate.is_valid_cert(self.quorum_membership.as_ref()) {
                     error!(
@@ -688,7 +674,6 @@ impl<TYPES: NodeType, I: NodeImplementation<TYPES>> ConsensusTaskState<TYPES, I>
                     };
                 }
             }
-            #[cfg(not(feature = "dependency-tasks"))]
             HotShotEvent::QuorumVoteSend(vote) => {
                 let Some(proposal) = self.current_proposal.clone() else {
                     return;
