+++ crates/task-impls/src/quorum_proposal_recv/handlers.rs
@@ -1,4 +1,5 @@
 #![allow(dead_code)]
+#![cfg(feature = "dependency-tasks")]
 
 use std::sync::Arc;
 
@@ -23,14 +24,11 @@ use tracing::{debug, error, warn};
 
 use super::QuorumProposalRecvTaskState;
 use crate::{
-    consensus::{
-        helpers::{
-            fetch_proposal, validate_proposal_safety_and_liveness, validate_proposal_view_and_certs,
-        },
-        view_change::{update_view, SEND_VIEW_CHANGE_EVENT},
-    },
     events::HotShotEvent,
-    helpers::broadcast_event,
+    helpers::{
+        broadcast_event, fetch_proposal, update_view, validate_proposal_safety_and_liveness,
+        validate_proposal_view_and_certs, SEND_VIEW_CHANGE_EVENT,
+    },
 };
 
 /// Whether the proposal contained in `QuorumProposalRecv` is fully validated or only the liveness
