+++ crates/testing/src/predicates/upgrade.rs
@@ -1,3 +1,5 @@
+#![cfg(not(feature = "dependency-tasks"))]
+
 use std::sync::Arc;
 
 use async_trait::async_trait;
